+++ libraries/resources/api300/c7000/ethernet_network_provider.rb
@@ -9,23 +9,12 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-my_client = {
-  url: ENV['ONEVIEWSDK_URL'],
-  user: ENV['ONEVIEWSDK_USER'],
-  password: ENV['ONEVIEWSDK_PASSWORD']
-}
-
-# Example: Adds an unmanaged device (default action)
-oneview_unmanaged_device 'UnmanagedDevice1' do
-  client my_client
-  data(
-    model: 'Procurve 4200VL',
-    deviceType: 'Server'
-  )
-end
-
-# Example: Removes an unmanaged device
-oneview_unmanaged_device 'UnmanagedDevice1' do
-  client my_client
-  action :remove
+module OneviewCookbook
+  module API300
+    module C7000
+      # EthernetNetworkProvider
+      class EthernetNetworkProvider < OneviewCookbook::API200::EthernetNetworkProvider
+      end
+    end
+  end
 end