+++ resources/ethernet_network.rb
@@ -13,44 +13,26 @@ OneviewCookbook::ResourceBaseProperties.load(self)
 
 default_action :create
 
-action_class do
-  include OneviewCookbook::Helper
-  include OneviewCookbook::ResourceBase
-
-  def update_connection_template(item, bandwidth)
-    bandwidth = convert_keys(bandwidth, :to_s)
-    connection_template = OneviewSDK::ConnectionTemplate.new(item.client, uri: item['connectionTemplateUri'])
-    connection_template.retrieve!
-    if connection_template.like? bandwidth
-      Chef::Log.info("#{resource_name} '#{name}' connection template is up to date")
-    else
-      converge_by "Update #{resource_name} '#{name}' connection template settings" do
-        connection_template.update(bandwidth)
-      end
-    end
-  end
-end
-
 action :create do
-  item = load_resource
-  bandwidth = item.data.delete('bandwidth')
-  create_or_update(item)
-  update_connection_template(item, bandwidth: bandwidth) if bandwidth
+  klass = OneviewCookbook::Helper.get_resource_class(self, :EthernetNetwork)
+  res = klass.new(self)
+  res.create_or_update
 end
 
 action :create_if_missing do
-  item = load_resource
-  bandwidth = item.data.delete('bandwidth')
-  created = create_if_missing(item)
-  update_connection_template(item, bandwidth: bandwidth) if bandwidth && created
+  klass = OneviewCookbook::Helper.get_resource_class(self, :EthernetNetwork)
+  res = klass.new(self)
+  res.create_if_missing
 end
 
 action :reset_connection_template do
-  item = load_resource
-  item.retrieve!
-  update_connection_template(item, bandwidth: OneviewSDK::ConnectionTemplate.get_default(item.client)['bandwidth'])
+  klass = OneviewCookbook::Helper.get_resource_class(self, :EthernetNetwork)
+  res = klass.new(self)
+  res.reset_connection_template
 end
 
 action :delete do
-  delete
+  klass = OneviewCookbook::Helper.get_resource_class(self, :EthernetNetwork)
+  res = klass.new(self)
+  res.delete
 end