+++ spec/fixtures/cookbooks/oneview_test/recipes/volume_create.rb
@@ -25,5 +25,4 @@ oneview_volume 'VOL1' do
   storage_system 'StorageSystem1'
   storage_pool 'Pool1'
   snapshot_pool 'Pool2'
-  volume_template 'Template1'
 end