+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 2.0.0
   - Support oneview-sdk v3.0 & different API versions/modules
 
+### 1.1.0
+  - Add support for client ENV variables
+  - Fixed volume resource (#92) & examples
+
 ## 1.0.0
   - Added support to Volume actions `:create_snapshot` and `:delete_snapshot`
   - Added support to SAN manager actions